final2

+++ main.c
@@ -5,6 +5,8 @@
 
 #include <math.h>
 #include <stdio.h>
+#include <math.h>
+
 
 // Usage:
 // Pass -DUSE_DOUBLE to the compiler to use double precision.
@@ -117,7 +119,7 @@ void rvv_add(fp_t *a, fp_t *b, fp_t *result, int size) {
     req_vlen = -1;
     asm volatile("vsetvli %0, %1, "vtype", m1 "tail_mask" " : "=r"(max_vlen) : "r"(req_vlen));
     req_vlen = size; // size in bits
-    int lmul = ceil((float)req_vlen / (float)max_vlen);
+    int lmul = (int)ceil((float)req_vlen / (float)max_vlen);
     if (lmul <= 1)
     {
